+++ src/pato/__init__.py
@@ -1,5 +1,5 @@
 from .utils.config import Config
 
-__version__ = "0.10.0"
+__version__ = "1.2.0"
 
 Config.set()